Only systems with snap pre-installed use /snap as root dir
The last iteration of this function presumed that the snap installed
executables would resolve to /snap/something, but this is not the case
for all Linux systems.

For Fedora and Arch the bin directory is currently set to
/var/lib/snapd/snap and when installing flutter you have to symlink
this dir to /snap. But since it is a symlink executable-find will
still resolve to the proper path which is "/var/lib/snapd/snap/flutter"
and not "/snap/flutter".

This patch is a little more lax on what constitues a snap install by
searching for the word "snap" in the file structure in order to
hopefully catch most of these variants.